truck: move redis to common folder
