+++ src/modules/user/user.module.ts
@@ -9,7 +9,7 @@ import { DeleteUserService } from './services/delete-user.service';
 import { GetUserByFilterService } from './services/user-by-filter.service';
 import { PasswordService } from './services/password.service';
 import { EmailService } from './services/email.service';
-import { RedisCacheService } from '../auth/infra/redis/redis-cache.service';
+import { RedisCacheService } from '../../common/redis/redis-cache.service';
 
 @Module({
   controllers: [UserController],
